requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
langchain==0.2.1
ollama
langchain-openai
replicate
python-dotenv==1.0.1
fastapi==0.110.3
pydantic==2.6.2
pytest==8.2.1
httpx==0.25.2
starlette==0.37.2